Add deactivateDriver method to DriverController and DriverService
This commit adds a new method called `deactivateDriver` to the `DriverController` and `DriverService` classes. The method allows deactivating a driver by setting their status to "Inactivo". The method takes a `Driver` object as a parameter and returns a response with a success message if the driver is successfully deactivated. If an error occurs, an appropriate error message is returned.